backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
yfinance==0.2.36
requests==2.31.0
beautifulsoup4==4.12.3

python-dotenv==1.0.1
werkzeug==3.0.1
pandas==2.1.4
websockets>=13.0
